pyproject.toml
[project]
name = "smtpy"
version = "0.1.0"
description = "Addy.io clone for custom email aliasing and forwarding."
authors = [
    {name = "SMTPy Team", email = "team@smtpy.local"}
]
requires-python = ">=3.9"

dependencies = [
    "aiosmtpd~=1.4.6",
    "fastapi~=0.116.0",
    "jinja2~=3.1.4",
    "dkimpy~=1.1.8",
    "uvicorn~=0.32.1",
    "httpx~=0.28.1",
    "python-multipart~=0.0.17",
    "passlib[bcrypt]~=1.7.4",
    "itsdangerous~=2.2.0",
    "dnspython~=2.7.0",
    "stripe~=12.3.0",
    "SQLAlchemy~=2.0.41",
    "starlette~=0.46.2",
    "pydantic-settings~=2.10.1",
    "aiosqlite~=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest~=8.4.1",
    "pytest-asyncio~=0.24.0",
    "pytest-cov~=6.0.0",
    "httpx~=0.28.1",
    "black~=24.10.0",
    "ruff~=0.8.4",
    "mypy~=1.13.0",
    "pre-commit~=4.0.1",
]

[tool.uv]
# uv-specific settings can go here

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.9"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "aiosmtpd.*",
    "dkimpy.*",
    "dns.*",
    "stripe.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "venv/*",
    ".venv/*",
    "*/migrations/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.hatch.build.targets.wheel]
packages = [
    "database",
    "views", 
    "controllers",
    "services",
    "utils",
    "smtp_server",
    "forwarding",
    "config_dns",
    "templates",
    "static"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]
